Fire and forget most voctomix commands
This is a remenant of the first prototype, using AJAX requests that
needed to provide a state-update response, rather than a websocket that
will provide asynchoronous updates as they occur.

We only need to wait for the response during connection when we're
building up our connection's state dictionary. At run-time we'll deliver
state updates over the websocket as we receive them.